so i did something stupid but importing players kinda works now
So like the game will prompt you after you have started and asked you if you want to play with the same amount of players you played with last time. And this surpasses the bug of there being a different number of players when you are playing a different game.